% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spfi_functions.R
\name{adjustAlaska}
\alias{adjustAlaska}
\title{(SPFI) Sum SEAK CWT data}
\usage{
adjustAlaska(x, data.catch)
}
\arguments{
\item{x}{A data frame, representation of the CWT data. Equivalent to a subset
of the hrj data frame where data.type=="NomCat"}

\item{data.catch}{A list with structure equivalent to the output from the function \code{\link{readCatchData}}.}
}
\value{
The function returns a data frame representation of the CWT data, in
  same format as the first input argument.
}
\description{
CWT data in SEAK fishery 4 is revised to equal sum of fishery
  4+6. This is a summation that is unique to SPFI estimation for Alaska only.
  If estimating for Alaska, then the hrj database variables: NomCat# (VB
  variable:cwtcatch), AEQCat# (VB variable:aeqcwtcat), and AEQTot# (VB
  variable:aeqcwttotmort) for fishery 4 are revised to equal sum of fishery
  4+6. The basis of this summation is not documented, but it may be partly
  associated with the fact that fishery 6 doesn't exist in the catch data
  file. These are the fishery definitions:
  \tabular{ccccccc}{ fishery \tab gear \tab fishery \tab fishery \tab fishery
  \tab fishery \tab fishery \cr index   \tab      \tab type    \tab country
  \tab name    \tab region \tab  psc\cr 4 \tab T \tab P \tab US \tab AK JLO T
  \tab AK \tab AABM\cr 6 \tab T \tab P \tab US \tab AK FALL T \tab AK \tab
  AABM }
}
\details{

}
\examples{
\dontrun{
#user needs to define catch file:
data.catch <- readCatchData(...)
#this assumes hrj.df has been read in and transformed to long format:
cwtcatch <- hrj.df[hrj.df$data.type=="NomCat"
                   & hrj.df$fishery.index \%in\% 1:6
                   & hrj.df$stock.index \%in\% C(1,9,10,15,17,20),]
cwtcatch <- adjustAlaska(x = cwtcatch, data.catch = data.catch)
}

}

