% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{build_jurisdiction}
\alias{build_jurisdiction}
\title{Create new jurisdiction data frame.}
\usage{
build_jurisdiction(stocfile.names)
}
\arguments{
\item{stocfile.names}{A character vector of *.STF files.}
}
\value{
A data frame comprising unique row for each stock in the *.STF files
  and those found in the jurisdiction.rda file.
}
\description{
Create new jurisdiction data frame.
}
\details{
This combines all stocks found in the stocfiles listed along with
  the currently available jurisdiction data frame. The user can then see what
  stocks are missing from the jurisdiction data frame and update the country
  and jurisdiction columns for those new stocks.
}
\examples{
\dontrun{
path <- "."
stocfile.names <- list.files(path = path, pattern = "*.STF$", full.names = TRUE)
jurisdiction <- build_jurisdiction(stocfile.names)
View(jurisdiction)
jurisdiction[is.na(jurisdiction$stock.country),]
}
}
