% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{build_stocks.key}
\alias{build_stocks.key}
\title{Create a new stocks.key data frame from *.STF files.}
\usage{
build_stocks.key(stocfile.names, overwriteRDA = FALSE)
}
\arguments{
\item{stocfile.names}{A character vector of *.STF files.}

\item{overwriteRDA}{A logical, with default FALSE. If set to TRUE the current
stock_key.rda will be overwritten.}
}
\value{
A data frame of the stocks found in the *.STF files. If \code{overwriteRDA} is TRUE then the stocks_key.rda file is overwritten.
}
\description{
Create a new stocks.key data frame from *.STF files.
}
\examples{
\dontrun{
path <- "."
stocfile.names <- list.files(path = path, pattern = "*.STF$", full.names = TRUE)
stocks.key <- build_stocks.key(stocfile.names)
}
}
