% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isbm_functions.R
\name{buildCMZfishery}
\alias{buildCMZfishery}
\title{(ISBM/CYER) Build a mapping table to define a fishery group's country.}
\usage{
buildCMZfishery(mort.df, fishery.list = list(list(country = "ca", fishery.term
  = c("nbc", "cbc", "wcvi", "bc", "canada")), list(country = "us", fishery.term
  = c("seak", "falcon", "wac", "puget", "us"))))
}
\arguments{
\item{mort.df}{A data frame. The element named \code{data.mort.long}, which
is in the output from \code{\link{readMortalityDist}}.}

\item{fisheries.list}{A list containing one or more lists. Each sub-list will
have the elements \code{country} and \code{fishery.term}. The
\code{fishery.term} is a character string to search for in the column
\code{mort.df$fisherygroup}.}
}
\value{
A data frame.
}
\examples{
\dontrun{
data.path <- "../data/totalmortalityDistribution"
mort.filenames <- list.files(data.path, pattern = ".csv$")
mort.filepaths <- paste(data.path, mort.filenames, sep="/")
mort.list <- readMortalityDist(mort.filepaths)
mort.df <- mort.list$data.mort.long
buildCMZfishery(mort.df=mort.df)
}
}
