% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spfi_functions.R
\name{buildSPFIscript}
\alias{buildSPFIscript}
\title{(SPFI) Build, save, and open an R script to help execute SPFI.}
\usage{
buildSPFIscript()
}
\description{
This creates and opens a script named "spfi_script.R". This is a
  template for the user to work with when doing SPFI estimates. This is
  intended to help the user understand the work flow and due to file path
  differences, is unlikely to work as is. Some object values will need
  updating (for example the datapath). If the user does not have a copy of
  the hrj data base saved  into a .Rdata file, then the functions
  \code{\link{readHRJAccessDatabase}}, \code{\link{reshapeHRJ}}, and the \code{list}
  creation will have to executed. It might be wise to then save the
  \code{list} to a .Rdata file.
}
\examples{
buildSPFIscript()
}

