% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spfi_functions.R
\name{calc_AAPC}
\alias{calc_AAPC}
\title{(SPFI) Data imputation by the Accrued APC (AAPC) method.}
\usage{
calc_AAPC(data.df, data.catch, stratum.var = "fishery.index",
  year.var = "return.year", value.var = "N.ty")
}
\arguments{
\item{data.df}{A data frame. Typically output from \code{\link{calc_N.ty}}.}

\item{data.catch}{A list with structure equivalent to the output from the
function \code{\link{readCatchData}}.}

\item{stratum.var}{A string. The name of the stratum variable. Default is
"fishery.index".}

\item{year.var}{A string. The name of the year variable. Default is
"return.year".}

\item{value.var}{A string. The name of the data variable. Default is "N.ty".}
}
\value{
A list of four elements named: \code{model,
  imputation.results, annual.estimate, data.df}. The first, \code{model}, is
  a list of the linear regression results for each stratum (including the
  data used in fit), the second, \code{imputation.results}, is a data frame
  of two columns. The first column usually has the same name as the
  \code{year.var} argument and the second is \code{N.t}. The other data
  frames represent the values calculated during intermediary steps. The time
  series of proportions for year with complete strata can be found in
  data.df.
}
\description{
This is similar to APC imputation as described on page 99 and
  Appendix 5 of REPORT TCCHINOOK (09)-2
  (\url{http://www.psc.org/download/35/chinook-technical-committee/2120/tcchinook09-2.pdf}).
   The abundance from a year-stratum contributes to the mean proportion
  estimate if the year-stratum has catch >1000.
}

