% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spfi_functions.R
\name{calc_Difference}
\alias{calc_Difference}
\title{(SPFI) Calculate difference between distribution parameters after each
  iteration.}
\usage{
calc_Difference(d.tsa.prior, d.tsa)
}
\arguments{
\item{d.tsa.prior}{A data frame. This is a copy of \code{d.tsa}, made before
an iteration loop during which \code{d.tsa} will be revised.}

\item{d.tsa}{A data frame. Output of \code{\link{calc_d.tsa}}.}
}
\value{
A vector of length one.
}
\description{
The user is not typically going to use this function. It is
  called by \code{\link{calc_SPFI}}.
}

