% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spfi_functions.R
\name{calc_GLMC}
\alias{calc_GLMC}
\title{(SPFI) Data imputation by the GLMC method.}
\usage{
calc_GLMC(data.df, data.catch, stratum.var = "fishery.index",
  year.var = "return.year", value.var = "N.ty", catchmin = 0)
}
\arguments{
\item{data.df}{A data frame. Typically output from \code{\link{calc_N.ty}}.}

\item{data.catch}{A list with structure equivalent to the output from the
function \code{\link{readCatchData}}.}

\item{stratum.var}{A string. The name of the stratum variable. Default is
"fishery.index".}

\item{year.var}{A string. The name of the year variable. Default is
"return.year".}

\item{value.var}{A string. The name of the data variable. Default is "N.ty".}

\item{catchmin}{An integer of length one. The minimum catch criteria for
inclusion of a stratum in the estimation. Strata below this value are set
to NA and estimated by imputation. Default is 0 (i.e. no strata are reset).}
}
\value{
A list of four data frames. The data frames are:
  \code{imputation.results, annual.estimate, prop.mean, data.df}. The first,
  \code{imputation.results}, comprises two columns. The first column usually
  has the same name as the \code{year.var} argument and the second is
  \code{N.t}. The other data frames represent the values calculated during
  intermediary steps. The time series of proportions for year with complete
  strata can be found in data.df.
}
\description{
This reproduces the results of the GLMC imputation as described by John Carlile.
}
\examples{
\dontrun{
N.ty <- calc_N.ty(T.ty = T.ty, hcwt.ty = hcwt.ty)
results.list <- calc_GLMC(N.ty, data.catch)
}
}
