% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spfi_functions.R
\name{calc_H.ty}
\alias{calc_H.ty}
\title{(SPFI) Calculate the AEQ stratum harvest rate parameters grouped by
  fishery, and year.}
\usage{
calc_H.ty(c.ty.sum, r.ty.sum, hcwt.ty)
}
\arguments{
\item{c.ty.sum}{Output from \code{\link{calc_ty.sum}}.}

\item{r.ty.sum}{Output from \code{\link{calc_ty.sum}}.}

\item{hcwt.ty}{Output from \code{\link{calc_hcwt.ty}}.}
}
\description{
This is equivalent to equation 6 in the draft SPFI document. In
  the Visual Basic this is referred to as the AEQScaler.
}

