% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spfi_functions.R
\name{calc_H.ty2}
\alias{calc_H.ty2}
\title{(SPFI) Calculate the AEQ stratum harvest rate parameters grouped by
  fishery, and year.}
\usage{
calc_H.ty2(c.ty.sum, r.ty.sum, hcwt.ty, T.ty)
}
\arguments{
\item{c.ty.sum}{Output from \code{\link{calc_ty.sum}}.}

\item{r.ty.sum}{Output from \code{\link{calc_ty.sum}}.}

\item{hcwt.ty}{Output from \code{\link{calc_hcwt.ty}}.}

\item{T.ty}{Output from \code{\link{calc_T.ty}}.}
}
\value{
A data frame of the AEQ stratum harvest rate parameter estimates
  grouped by fishery stratum and year.
}
\description{
This is equivalent to equation 6a in the draft SPFI document. In
  the Visual Basic this is referred to as the AEQScaler.
}
\examples{
\dontrun{
T.ty <- calc_T.ty(catch.df = data.catch$data.catch)
H.ty <- calc_H.ty2(c.ty.sum = c.ty.sum, r.ty.sum = r.ty.sum, hcwt.ty = hcwt.ty, T.ty = T.ty)
}
}
