% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spfi_functions.R
\name{calc_H.ty3}
\alias{calc_H.ty3}
\title{(SPFI) Calculate the AEQ stratum harvest rate parameters grouped by
  fishery, and year.}
\usage{
calc_H.ty3(c.ty.sum, r.ty.sum, T.ty, N.ty)
}
\arguments{
\item{c.ty.sum}{Output from \code{calc_ty.sum}}

\item{r.ty.sum}{Output from \code{calc_ty.sum}}

\item{T.ty}{Output from \code{calc_T.ty}.}
}
\description{
This is equivalent to equation 6b in the draft SPFI document. In
the Visual Basic this is referred to as the AEQScaler.
}

