% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spfi_functions.R
\name{calc_H.y}
\alias{calc_H.y}
\title{(SPFI) Calculate the AEQ fishery harvest rate parameters grouped by
  year.}
\usage{
calc_H.y(c.ty.sum, r.ty.sum, hcwt.ty, T.ty)
}
\arguments{
\item{c.ty.sum}{Output from \code{\link{calc_ty.sum}}.}

\item{r.ty.sum}{Output from \code{\link{calc_ty.sum}}.}

\item{hcwt.ty}{Output from \code{\link{calc_hcwt.ty}}.}

\item{T.ty}{Output from \code{\link{calc_T.ty}}.}
}
\value{
A data frame of the AEQ fishery harvest rate parameter estimates
  grouped by year.
}
\description{
This is equivalent to equation 7 in the draft SPFI document.
}
\examples{
\dontrun{
H.y <- calc_H.y(c.ty.sum = c.ty.sum, r.ty.sum = r.ty.sum, hcwt.ty = hcwt.ty, T.ty = T.ty)
}
}
