% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spfi_functions.R
\name{calc_H.y2}
\alias{calc_H.y2}
\title{(SPFI) Calculate the harvest rate parameters grouped by year.}
\usage{
calc_H.y2(c.ty.sum, r.ty.sum, T.ty, N.y)
}
\arguments{
\item{c.ty.sum}{Output from \code{\link{calc_ty.sum}}.}

\item{r.ty.sum}{Output from \code{\link{calc_ty.sum}}.}

\item{T.ty}{Output from \code{\link{calc_T.ty}}.}

\item{N.y}{Output from \code{\link{calc_N.y}}.}
}
\description{
This is equivalent to equation 7b in the draft SPFI document.
This function produces the same output as \code{\link{calc_H.y}}.
}

