% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spfi_functions.R
\name{calc_hcwt.ty}
\alias{calc_hcwt.ty}
\title{(SPFI) Calculate the CWT harvest rate parameters grouped by fishery,
  and year.}
\usage{
calc_hcwt.ty(r.ty.sum, d.tsa, n.ysa)
}
\arguments{
\item{r.ty.sum}{Output from \code{\link{calc_ty.sum}}.}

\item{d.tsa}{Output from \code{\link{calc_d.tsa}}.}

\item{n.ysa}{Synonymous with \code{CWTPop} in VB or:
\code{hrj.df[hrj.df$data.type=="Pop" & hrj.df$fishery.index == 1,]}}
}
\value{
A data frame of the harvest rate parameter estimates grouped by
  fishery stratum and year.
}
\description{
This is equivalent to equation 2 in the draft SPFI document.
  This function does a single estimate of the harvest rate values. It does
  not perform iterations to optimize.
}
\examples{
\dontrun{
#look to \\code{\\link{buildSPFIscript}} for creating hrj.df
hrj.df <- hrj.df[hrj.df$spfiflag==1,]
cwtpop <- hrj.df[hrj.df$data.type=="Pop" & hrj.df$fishery.index == 1 & hrj.df$stock.index \%in\% stock.subset,]
cwtpop <- subset(cwtpop,select = -fishery.index) #n.ysa
cwtcatch <- hrj.df[hrj.df$data.type=="NomCat" & hrj.df$fishery.index \%in\% fishery.subset & hrj.df$stock.index \%in\% stock.subset,]
if(region=="seak") cwtcatch <- adjustAlaska(x = cwtcatch, data.catch = data.catch)
r.ty.sum <- calc_ty.sum(x = cwtcatch, newvar.name = "r.ty.sum")
d.tsa <- calc_d.tsa(r.tsa.sum = r.tsa.sum, n.ysa = cwtpop, standardize.bol = TRUE)
hcwt.ty <- calc_hcwt.ty(r.ty.sum=r.ty.sum,  d.tsa = d.tsa, n.ysa = cwtpop)
}
}
