% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spfi_functions.R
\name{calc_N.ty}
\alias{calc_N.ty}
\title{(SPFI) Calculate the abundance by fishery stratum and year.}
\usage{
calc_N.ty(T.ty, hcwt.ty)
}
\arguments{
\item{T.ty}{Output from \code{\link{calc_T.ty}}.}

\item{hcwt.ty}{Output from \code{\link{calc_hcwt.ty}}.}
}
\description{
This is equivalent to equation 4 in the draft SPFI document.
}
\examples{
\dontrun{
N.ty <- calc_N.ty(T.ty = T.ty, hcwt.ty = hcwt.ty)
}
}

