% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spfi_functions.R
\name{calc_N.y}
\alias{calc_N.y}
\title{(SPFI) Calculate the total yearly abundance across fishery strata.}
\usage{
calc_N.y(N.ty)
}
\arguments{
\item{N.ty}{Output from \code{\link{calc_N.ty}}.}
}
\value{
A data frame of the abundance estimates, grouped by year.
}
\description{
This is equivalent to equation 5 in the draft SPFI document.
}
\examples{
\dontrun{
T.ty <- calc_T.ty(catch.df = data.catch$data.catch)
N.ty <- calc_N.ty(T.ty = T.ty, hcwt.ty = hcwt.ty)
N.y <- calc_N.y(N.ty = N.ty)
}
}
