% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spfi_functions.R
\name{calc_S.ty}
\alias{calc_S.ty}
\title{(SPFI) Calculate the stata specific harvest rate indices by year.}
\usage{
calc_S.ty(H.ty)
}
\arguments{
\item{H.ty}{Output from \code{calc_H.ty} or \code{calc_H.ty2} or \code{calc_H.ty3}.}
}
\description{
This is equivalent to equation 8 in the draft SPFI document.
}

