% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spfi_functions.R
\name{calc_S.y}
\alias{calc_S.y}
\title{(SPFI) Calculate the SPFI grouped by year.}
\usage{
calc_S.y(H.y)
}
\arguments{
\item{H.y}{Output from \code{\link{calc_H.y}} or \code{\link{calc_H.y2}}.}
}
\value{
A data frame with the SPFI values by year.
}
\description{
This is equivalent to equation 9 in the draft SPFI document.
}
\examples{
\dontrun{
S.y <- calc_S.y(H.y = H.y)
}
}
