% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spfi_functions.R
\name{calc_T.ty}
\alias{calc_T.ty}
\title{(SPFI) Calculate the Pacific Salmon Treaty catch, grouped by fishery
  stratum and year.}
\usage{
calc_T.ty(catch.df)
}
\arguments{
\item{catch.df}{A data frame, extracted from the output of
\code{\link{readCatchData}}, such as: \code{data.catch$data.catch}.}
}
\value{
A data frame.
}
\description{
This is equivalent to equation 3 in the draft SPFI document. The
  output is the same as found in the Visual Basic variable:
  \code{CatchContribution}. The SEAK data includes values for "Alaska
  hatchery contribution". The "Alaska hatchery contribution" is subtrated
  from the catch (by stratum and year) to produce the PSC catch.
}
\examples{
\dontrun{
data.catch <- readCatchData("wcvi7914.cat", strLocation = 'wcvi')
T.ty <- calc_T.ty(catch.df = data.catch$data.catch)
}
}
