% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spfi_functions.R
\name{calc_tsa.sum}
\alias{calc_tsa.sum}
\title{(SPFI) Sum CWT catch by fishery, stock, and age.}
\usage{
calc_tsa.sum(x, newvar.name = "value.sum")
}
\arguments{
\item{x}{A data frame. Typically the \code{cwtcatch} object.}

\item{newvar.name}{A string of length one. This is the new name for the
column of summed data. The default is "value.sum". It's recommended not to
change the default unless using this function for purposes independent of
the SPFI estimation.}
}
\value{
A data frame of the CWT catch sum grouped by fishery, stock, and age.
}
\examples{
\dontrun{
  cwtcatch <- hrj.df[hrj.df$data.type=="NomCat" & hrj.df$fishery.index \%in\% fishery.subset & hrj.df$stock.index \%in\% stock.subset,]
if(region=="seak") cwtcatch <- adjustAlaska(x = cwtcatch, data.catch = data.catch)
r.tsa.sum <- calc_tsa.sum(x = cwtcatch, newvar.name = "r.tsa.sum")
}
}
