% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isbm_functions.R
\name{calcCYER}
\alias{calcCYER}
\title{(ISBM/CYER) Calculate CYER and its index.}
\usage{
calcCYER(mort.df, fishery.map, country = c("canada", "us"))
}
\arguments{
\item{mort.df}{A data frame. The element named \code{data.mort.long}, which
is in the output from \code{\link{readMortalityDist}}.}

\item{fishery.map}{A data frame. Output from \code{\link{buildCMZfishery}}.}
}
\value{
A data frame with the estimates of catch year exploitation rate and
  index.
}
\description{
(ISBM/CYER) Calculate CYER and its index.
}
\examples{
\dontrun{
data.path <- "../data/totalmortalityDistribution/qui_2-6"
mort.filenames <- list.files(data.path, pattern = ".csv$")
mort.filepaths <- paste(data.path, mort.filenames, sep="/")
mort.list <- readMortalityDist(mort.filepaths)
mort.df <- mort.list$data.mort.long
mort.df$year <- mort.df$CatchYear
fishery.map <- buildCMZfishery(mort.df = mort.df)
calcCYER(mort.df, fishery.map)
}
}
