% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isbm_functions.R
\name{calcCYERindex}
\alias{calcCYERindex}
\title{Calculate the Catch year exploitation rate index (CYER index).}
\usage{
calcCYERindex(cyer.df)
}
\arguments{
\item{cyer.df}{A data frame. Output from \code{\link{calcCYER}}}
}
\value{
A data frame with same structure as the argument \code{cyer.df} with
  the addition of a column: \code{cyer.index}.
}
\description{
Calculate the Catch year exploitation rate index (CYER index).
}
\examples{
\dontrun{
calcCYERindex(cyer.df)
}
}
