% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isbm_functions.R
\name{calcIndexError}
\alias{calcIndexError}
\title{Calculate errors of ISBM and CYER}
\usage{
calcIndexError(dat, postseason = 0:3, minyear = 2000, maxyear = 2013,
  estimationyear = 2016, pm = c("mpe", "mape"), standardize = FALSE, ...)
}
\arguments{
\item{dat}{A data frame. Typically the output from \code{\link{calcCYER}} or
\code{\link{combinePTdata}}.}

\item{postseason}{An integer vector. Defines what postseason periods to
evaluate. Default is 0:3, which equates to postseason years 1:4.}

\item{minyear}{An integer of length one. The initial year in the time series
for evaluation.}

\item{maxyear}{An integer of length one. The final year in the time series
for evaluation.}

\item{estimationyear}{An integer of length one. The final year in the time
series.}

\item{pm}{A character vector. Names of the performance measures to calculate.
Default tests are \code{mpe} and \code{mape}. Look to
\code{\link{PBSperformance::PMs}} for details.}
}
\value{
A list of two data frames. One data frame holds the annual error
  estimates, the other data frame has the performance measure statistic.
}
\examples{
\dontrun{
mort.list <- readMortalityDist("QUI2000_2-6_CMZ.csv")
mort.df <- mort.list$data.mort.long
mort.df$year <- mort.df$CatchYear
fishery.map <- buildCMZfishery(mort.df = mort.df)
dat.cyer <- calcCYER(mort.df = mort.df, fishery.map=fishery.map)
dat.cyer$index <- dat.cyer$mortality.prop/100
results.cyer <- calcIndexError(dat = dat.cyer)
}
}
