% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chinookmodelfunctions.R
\name{calcPMs}
\alias{calcPMs}
\title{(Calibration Performance) Group Performance Measures}
\usage{
calcPMs(data.combined, datasubset = c("escapement", "terminalrun"),
  pm = list(PBSperformance::mre, PBSperformance::mae, PBSperformance::mpe,
  PBSperformance::mape, PBSperformance::rmse), samplesize.min = 10,
  writecsv = TRUE, results.path = ".")
}
\arguments{
\item{data.combined}{A dataframe. Output of \code{\link{importFCSCCC}}.}

\item{datasubset}{A character vector defining the values from data.type to be selected.}

\item{pm}{A list of performance measures to calculate.}

\item{writecsv}{A Boolean confirming if output should also be written to a csv file.}
}
\value{
The function returns a dataframe comprising the performance metrics requested,
grouped by \code{calibration, stock, data.type, and agegroup}.
}
\description{
A wrapper to calculate multiple performance measures.
}
\details{
This is a convenient wrapper that estimates one or more performance
measures for the data supplied. This function is somewhat specialized for use
on CTC calibration model comparisons (most often against values in the *.fcs files.
The output from the \code{\link{mergeFCSCCC}} function is what normally would be used for
the argument \code{data.combined}.
This function expects \code{data.combined} to include columns with names:
\code{calibration, stock, data.type, agegroup, value.fcs, value.ccc}.
The latter two columns are, respectively, the data that will be used in the arguments:
\code{expect} and \code{obs} of the various performance metric functions.
}
\examples{
metrics <- calcPMs(data.combined, pm = list(PBSperformance::mpe,  PBSperformance::mape), writecsv = TRUE, samplesize.min = samplesize.min, results.path = model.list$results.path)

}
