% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chinookmodelfunctions.R
\name{calcRanks}
\alias{calcRanks}
\title{(Calibration Performance) Rank Calculation}
\usage{
calcRanks(dat, columnToRank, rank.method = c("ordinal", "interpolated"),
  abs.value = FALSE)
}
\arguments{
\item{dat}{A dataframe}

\item{columnToRank}{A vector of integers defining what columns to rank
(independently)}

\item{rank.method}{A character defining the ranking method to apply. The
method can be either 'ordinal' or 'interpolated'.}

\item{abs.value}{A Boolean vector with length equal to the length of the
columnToRank argument. This indicates whether or not to take the absolute
value of the data before the ranking. Default value is FALSE.}
}
\value{
A dataframe comprising all columns of the argument: \code{dat} and
  one additional column of ranks for each column column declared in
  \code{columnToRank}. The name of each rank column is a combination of the
  data column name and ".rank".
}
\description{
Calculate rank of data.
}
\examples{
\dontrun{
dat.test <- data.frame(id=1:20, values=rnorm(20))
calcRanks(dat.test,  columnToRank=2, abs.value = TRUE )
}
}
