% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newAWKfunctions.R
\name{checkMissingfiles}
\alias{checkMissingfiles}
\title{Check for matching B & C file names in the vector of file names.}
\usage{
checkMissingfiles(filepath)
}
\arguments{
\item{filepath}{A character vector, which can have length greater than one.
Each element of the vector is the path and filename for each HRJ file to be
read in.}
}
\description{
This tests for that every file named in the filepath, whether
  'b' or 'c' data, the alternate data ('c' or 'b') by stock, has also been
  named.
}

