% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isbm_functions.R
\name{cleanCMZ}
\alias{cleanCMZ}
\title{Clean *cmz.csv files for easy importing.}
\usage{
cleanCMZ(filenames)
}
\arguments{
\item{filenames}{A character vector of length one or more. The names of the
*CMZ.csv files.}
}
\value{
Nothing returned. The files listed in filenames are overwritten with
  the cleaned CMZ files
}
\examples{
\dontrun{
data.path <- '../data/'
mort.filenames <- list.files(data.path, pattern = '.csv$')
mort.filepaths <- paste(data.path, mort.filenames, sep='/')
cleanCMZ(mort.filepaths)
}
}
