% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{ctctools-deprecated}
\alias{ctctools-deprecated}
\alias{importData}
\alias{mergeData}
\alias{plotCompare}
\alias{writeTable1}
\alias{writeTable3}
\alias{writeTable4}
\alias{writeTable5}
\alias{write_table6.6}
\title{Deprecated functions in ctctools}
\usage{
importData(data.path.vec = NA, model.list = NULL)

mergeData(ccc.list, fcs.list, stocks.names = "all")

plotCompare(data.combined, savepng = FALSE, results.path = ".",
  point.col.df, ...)

writeTable1(data.combined, results.path = ".")

writeTable3(data.combined, results.path = ".")

writeTable4(data.combined, results.path = ".")

writeTable5(data.combined, results.path = ".")

write_table6.6(spfi.output, data.catch)
}
\description{
These functions still work but will be eventually be removed (defunct).
}
\details{
\itemize{
The deprecated functions include:
 \item \code{\link{importData}}
 \item \code{\link{mergeData}}
 \item \code{\link{writeTable1}}
 \item \code{\link{writeTable3}}
 \item \code{\link{writeTable4}}
 \item \code{\link{writeTable5}}
 \item \code{\link{write_table6.6}}

}
}

