% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{fishery.def}
\alias{fishery.def}
\title{Data defining each fishery.}
\format{A data frame with 78 rows and 7 variables:
\describe{
  \item{fishery.index}{The integer number of each fishery}
  \item{gear}{Troll = "T", net = "N", sport = "S", nonfishery = "X"}
  \item{fishery.type}{Preterminal = "P", terminal = "T"}
  \item{fishery.country}{"US" or "CA"}
  \item{fishery.name}{A slightly longer text string desribing fishery}
  \item{fishery.region}{}
  \item{fishery.psc}{}
   ... }}
\usage{
fishery.def
}
\description{
A dataset containing the fisheries and associated meta data.
}
\examples{
data(fishery.def)
#or
data("fishery.def")
str(fishery.def)
}
\keyword{datasets}

