% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chinookmodelfunctions.R
\name{importFCSCCC}
\alias{importFCSCCC}
\title{(Calibration Performance) Import FCS and CCC files}
\usage{
importFCSCCC(data.path.vec = NA, model.list = NULL)
}
\arguments{
\item{data.path.vec}{A vector of path names that point to data folders. This
argument is only needed if model.list is not included. And the code is
evolving such that it's best to include model.list. Eventually this
argument may be removed from the function.}

\item{model.list}{A list. Output of the buildmodel.list function.}
}
\value{
A data frame in long format, comprising both the FCS and CCC data.
}
\description{
Based on the criteria defined in the model.list, this function
  imports a single FCS file and one or more CCC files.
}

