% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chinookmodelfunctions.R
\name{importTableOfDifferences}
\alias{importTableOfDifferences}
\title{(Calibration Performance) Import and combine tables of differences
  from csv files.}
\usage{
importTableOfDifferences(ranking.method, pm.type.vec = c("mpe", "mape"),
  data.type = "escapement+terminalrun", results.path = ".")
}
\arguments{
\item{results.path}{}
}
\value{
A data frame in long format, that is combination of all csv files.
}
