% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{jurisdiction}
\alias{jurisdiction}
\title{Jurisdiction data for each stock}
\format{A data frame with 58 rows and 5 variables:
\describe{
  \item{stock}{three letter stock identifier}
  \item{stock.long}{the full stock name}
  \item{jurisdiction}{two letter identifier of province or state}
  \item{jurisdiction.long}{full name of province or state, in lower case}
  \item{stock.country}{two letter identifier country, but include 'TB' for transboundary}
   ... }}
\usage{
jurisdiction
}
\description{
A dataset containing the stocks and their associated jurisdictions, including
province or state.
}
\examples{
data(jurisdiction)
#or
data("jurisdiction")
str(jurisdiction)
}
\keyword{datasets}

