% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chinookmodelfunctions.R
\name{mergeData}
\alias{mergeData}
\title{(Calibration Performance) Merge FCS And CCC Data}
\usage{
mergeData(ccc.list, fcs.list, stocks.names = "all")
}
\arguments{
\item{ccc.list}{A list, see details}

\item{fcs.list}{A list, see details}
}
\value{
A dataframe is produced.
}
\description{
Merge FCS and CCC data
}
\details{
The function arguments are output from the functions \code{readCCC}
and \code{readFCS}, respectively.
}

