% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chinookmodelfunctions.R
\name{mergeFCSCCC}
\alias{mergeFCSCCC}
\title{(Calibration Performance) Merge FCS And CCC Data}
\usage{
mergeFCSCCC(ccc.list, fcs.list, stocks.names = "all")
}
\arguments{
\item{ccc.list}{A list, see details}

\item{fcs.list}{A list, see details}
}
\value{
A dataframe is produced.
}
\description{
Merge FCS and CCC data
}
\details{
The function arguments are output from the functions
  \code{\link{readCCC}} and \code{\link{readFCS}}, respectively. This
  function is called within \code{\link{importFCSCCC}} and the user is unlikely
  to use it on its own.
}
\examples{
\dontrun{
data.combined <- mergeFCSCCC(ccc.list = ccc.list, fcs.list = fcs.list)
}
}
