% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spfi_functions.R
\name{minimizeDistribution}
\alias{minimizeDistribution}
\title{Minimization to estimate distribution parameter}
\usage{
minimizeDistribution(data.type = c("AEQCat", "AEQTot"),
  region = c("wcvi", "nbc", "seak"), hrj.df = NA, hrj.filename = NA,
  data.catch, data.stock, fishery.subset = NULL, stock.subset = NULL,
  adjustAlaska.bol = TRUE)
}
\arguments{
\item{adjustAlaska.bol}{}
}
\description{
Minimization to estimate distribution parameter
}
\examples{
\dontrun{
data.type <- "AEQTot" # "AEQCat" # or "AEQTot"
region <- "seak" # "wcvi" # "nbc" #  "seak"
#only one aabm in a data folder:
data.catch <- readCatchData("wcvi7914.cat", strLocation = region)
data.stock <- readStockData( "STOCFILE.STF") load("hrj_from_mdb.RData")
#data.hrj is available from the load() above:
hrj.df <- data.hrj$hrj.list.long$HRJ_BY
#all data sets include data prior to 1979.
#These values are excluded in the VB.
year.range <- 1979:(data.stock$stockmeta$intLastBrood$value+1)
hrj.df <- hrj.df[hrj.df$return.year \%in\% year.range,]
minimizeDistribution(data.type = data.type, region = region, hrj.df = hrj.df,
data.catch = data.catch, data.stock = data.stock)
}
}
