% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp_functions.R
\name{plot_fpseries}
\alias{plot_fpseries}
\title{(FP) Plot time series of FP values.}
\usage{
plot_fpseries(dat.fp, savepng = FALSE, filename = NA)
}
\arguments{
\item{dat.fp}{A data frame. Output from \code{\link{calc_fp}}.}

\item{savepng}{A Boolean. Option to send output to a png file instead of plot
window. Default is FALSE.}
}
\value{
A lattice plot of the FP series, by stock and age. If \code{savepng}
  is TRUE then each stock is sent to an individual png file and the stock
  name is included in the file name.
}
\examples{
\dontrun{
plot_fpseries(dat.fp = dat.fp, savepng = TRUE)
}
}
