% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chinookmodelfunctions.R
\name{plotFCSCCC}
\alias{plotFCSCCC}
\title{(Calibration Performance) Model Comparison Plots}
\usage{
plotFCSCCC(data.combined, savepng = FALSE, results.path = ".", point.col.df,
  ...)
}
\arguments{
\item{data.combined}{A dataframe, see details}

\item{savepng}{A Boolean confirming to also save a PNG file of each plot.}

\item{...}{}

\item{agegroup.n}{An integer, equates to the number of plots per page. See
detail}
}
\value{
One lattice plot per unique combination of stock and calibration model,
with option to produce the same in the form of a PNG file.
}
\description{
1:1 plots for comparing FCS and CCC data by stock and age.
}
\details{
Option to produce PNG files. Relies on the package: \code{lattice}.
  Argument \code{data.combined} Argument \code{agegroun.n} determines the
  number of lattice panels per page. The value would typically range 1--4
  based on the number of age groups being evaluated. Considering the unique
  values of \code{agegroup} are c('age.3', 'age.4', 'age.5',
  'totalabundance').
}
