% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chinookmodelfunctions.R
\name{plotFCSvsCCC}
\alias{plotFCSvsCCC}
\title{plotFCSvsCCCfc}
\usage{
plotFCSvsCCC(data.combined, samplesize.min, results.path = ".",
  point.col.df = data.frame(year.start = c(1979, 1985, 1999, 2009), year.end =
  c(1984, 1998, 2008, NA), point.col = c("black", "yellow", "red", "green"),
  stringsAsFactors = FALSE), ...)
}
\arguments{
\item{...}{}
}
\description{
plotFCSvsCCCfc
}
