% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp_functions.R
\name{plotFPAseries}
\alias{plotFPAseries}
\title{Plot series found in *.fpa files.}
\usage{
plotFPAseries(dat.fpa, stock.var = "stockname_long", age.var = "age",
  plot.true.age = TRUE, aabm, fpa.filename)
}
\arguments{
\item{dat.fpa}{A data frame. Output from \code{\link{read_FPA}}.}

\item{stock.var}{A character vector of length one. The name of the variable
representing stock names.}

\item{age.var}{A character vector of length one. The name of the variable
representing age data (i.e. the age index or true age).}

\item{plot.true.age}{Logical (TRUE/FALSE). Plot true age data. Default is
TRUE. If TRUE, then \code{age.var} should likely be "age".}

\item{aabm}{A character vector of length one. The AABM name. This is used in
the ouput filename.}

\item{fpa.filename}{A character vector of length one. The FPA name. This is
used in the ouput filename.}
}
\value{
A single png file of lattice plot showing the FPA series by stock and
  age or age index.
}
\description{
Plot series found in *.fpa files.
}
\examples{
\dontrun{
plotFPAseries(dat.fpa = dat.fpa, stock.var = "stockname_long", age.var = "age",
 plot.true.age = TRUE, aabm = aabm, fpa.filename = fpa.filename)
}
}
