% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chinookmodelfunctions.R
\name{plotPM}
\alias{plotPM}
\title{(Calibration Performance) Lattice plot to compare PMs across
  calibration models}
\usage{
plotPM(tableofdifferences, results.path = ".", savepng = FALSE)
}
\arguments{
\item{tableofdifferences}{A data frame, Output of
\code{\link{importTableOfDifferences}}}

\item{results.path}{A character vector of length 1. Defines the path for the
"Table of differences" csv files. Default is the current working directory.}

\item{savepng}{A Boolean. Save output to a png file. Default is FALSE.}
}
\value{
Lattice plot of the data found in the "Table of differences" files.
}

