% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp_functions.R
\name{read_FPA}
\alias{read_FPA}
\title{Read FPA files.}
\usage{
read_FPA(filenames)
}
\arguments{
\item{filenames}{A character vector. The names of *.fpa files for importing.}
}
\value{
A list, the same length as the number of fpa files in
  \code{filenames}. Each list element is also a list and comprises of four
  elements: the filename, metadata (information in the first four rows of the
  file), and two data frames, the data in wide format similar to that found
  in the fpa file, and data in long format. The rows are defined by two row
  variables. The variable \code{model.stocknumber} is the value found in the first
  column of the fpa file (after row 4, the year row). \code{age.index} is
  a sequence of 1 to 4 indicating the subrow location below each "main" row.
}
\description{
This reads in one or more fpa files and exports the data in two
  formats. The first is similar to the structure found in the fpa file (wide
  format), the second is in long format, which is more useful for analysis in
  R.
}
\examples{
\dontrun{
filenames <- list.files(pattern = "*\\\\.fpa")
dat.out <- read_FPA(filenames)
}
}
