% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp_functions.R
\name{read_mdl}
\alias{read_mdl}
\title{Read in mdl files of cwt recovery data.}
\usage{
read_mdl(filenames)
}
\arguments{
\item{filenames}{A character vector of the mdl file names.}
}
\value{
A list of two elements. The first element, named \code{dat.mdl}, is also a list. Each element of \code{dat.mdl} is also a list and represents the data from one mdl file. The second element of the output list, named \code{dat.mdl.long}, is a data frame of all the combined mdl files. This latter list element is the preferred source of data for analysis.
}
\examples{
\dontrun{
mdl.filenames <- list.files(path = "../data/2015BPC_fpa_AABM_troll_spfi - 9-20-2016/56FMDL", pattern = "MDL$")
mdl.filepath <- paste("../data/2015BPC_fpa_AABM_troll_spfi - 9-20-2016/56FMDL", mdl.filenames, sep="/")
dat.mdl <- read_mdl(mdl.filepath)
fisheries.needed <- fishery.def[fishery.def$aabm \%in\% aabm,]
dat.mdl.long.sub <- dat.mdl$dat.mdl.long[dat.mdl$dat.mdl.long$fishery.name \%in\% fisheries.needed$fishery.name,]
}
}
