% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp_functions.R
\name{read_modelstocklist.csv}
\alias{read_modelstocklist.csv}
\title{Read model stock list file.}
\usage{
read_modelstocklist.csv(filename)
}
\arguments{
\item{filename}{A character vector of length one.}
}
\value{
A list with two elements. The first element is the meta-data found at
  the top of the file. The second element is a data frame comprising the data
  of stock names etc.
}
\description{
Read model stock list file.
}
\examples{
\dontrun{
model.stocklist <- read_modelstocklist.csv("NewModelStockList.csv")
}
}
