% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp_functions.R
\name{read_stkfile}
\alias{read_stkfile}
\title{Read in stk file of base period exploitation rates.}
\usage{
read_stkfile(filename, baseperiodfishery.names)
}
\arguments{
\item{filename}{A character vector of length 1. The stk file of base period
exploitation rates.}

\item{baseperiodfishery.names}{A character vector. Likely based on the file
"48FisheryName.txt".}
}
\value{
A list comprising two elements. The first, named \code{dat.meta.long}
  is a data frame of data other than the ER values. This includes the initial
  cohort abundance (\code{initcohortabun}), the maturation rates
  (\code{matrates}), and the AEQ factor (\code{aeqfactor}). The second list
  element is \code{dat.er.long}, which is a data frame of the ER data.
}
\examples{
\dontrun{
baseperiodfishery.names <- readLines(
   "../data/2015BPC_fpa_AABM_troll_spfi - 9-20-2016/48FisheryName.txt")
dat.stk <- read_stkfile(
  "../data/2015BPC_fpa_AABM_troll_spfi - 9-20-2016/2015BPC_PII_V1.5.STK",
   baseperiodfishery.names)
}
}
