% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chinookmodelfunctions.R
\name{readCCC}
\alias{readCCC}
\title{(Calibration Performance) Import CCC Files}
\usage{
readCCC(filepath, data.types = c("AEQ", "cohort", "termrun", "escape"),
  stocks.key = NA)
}
\arguments{
\item{filepath}{A character vector of the CCC files to be imported. See details}

\item{data.types}{See details}

\item{stocks.key}{See details}
}
\value{
A list, with one element per CCC (calibration) file.
}
\description{
Import one or more CCC files into a list.
}
\details{
Need details.
}
