% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readDataArchive}
\alias{readDataArchive}
\title{Read text files that are stored in a PBS archive format.}
\usage{
readDataArchive(filepaths.list)
}
\arguments{
\item{filepaths.list}{A list of individual filenames for importing.}
}
\value{
A list comprising any meta data included in the original text file
  and the data, which may be a vector or data frame.
}
\description{
Read text files that are stored in a PBS archive format.
}
\details{
This format is created by Michael Folkes (DFO) and not in
  circulation.
}
\examples{
\dontrun{
dat.tmp <- readDataArchive(filepaths.list)
}
}
