% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newAWKfunctions.R
\name{readHRJtext}
\alias{readHRJtext}
\title{Read in and combine multiple HRJ text files.}
\usage{
readHRJtext(filepath)
}
\arguments{
\item{filepath}{A character vector, which can have length greater than one.
Each element of the vector is the path and filename for each HRJ file to be
read in. As each file has its own path informartion, files can be read in
from multiple folders.}
}
\value{
A list comprising two data frames of the HRJ data in 'wide' format.
  The escapement data are found in a third data frame of the output list.
  The escapement data are in 'long' format, with structure matching the output of
  \code{\link{reshapeHRJ}}.
}
\description{
The function reads in multiple HRJ files and produces two data
  frames, one for 'b' data and one for the 'c' data. The format is wide,
  matching that found in the HRJ database. The escapement data is also
  extracted but tabulated separately.
}
\examples{
\dontrun{
filename <- list.files(".", pattern = "HRJ")
hrj.list.long <- readHRJtext(filename)
}
}

