% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isbm_functions.R
\name{readMortalityDist}
\alias{readMortalityDist}
\title{(ISBM/CYER) Import motality distribution CMZ (*.csv) files.}
\usage{
readMortalityDist(filenames, ...)
}
\arguments{
\item{filenames}{A character vector of length one or more. The names of the
*CMZ.csv files.}
}
\value{
A list comprising two data frames. The first data frame
  (\code{data.mort.wide}) has the same structure as found in the csv file.
  The second data frame (\code{data.mort.long}) is reshaped to a long format
  and the fishery column names are parsed into additional columns.
}
\examples{
\dontrun{
mort.list <- readMortalityDist(mort.filepaths)
mort.df <- mort.list$data.mort.long
View(mort.list[mort.df$fishery.psc=="ISBM",])
}
}
