% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isbm_functions.R
\name{readPTfromstep11csv}
\alias{readPTfromstep11csv}
\title{(ISBM/CYER) Import one or more ISBM index files, in step11 csv format.}
\usage{
readPTfromstep11csv(filenames, agerange = NA)
}
\arguments{
\item{filenames}{A character vector. The names of the *step11*.csv files for
import.}

\item{agerange}{A character vector of length one. This overrides the age
range value declared in the file name. e.g. "2-6"}
}
\value{
A list comprising elements that are also lists, one for each .csv
  file imported. Each sub list comprises six elements. Their names are:
  \code{metadata}, \code{data.isbm}, \code{data.isbm.long},
  \code{data.isbm.ages}, \code{data.bper}, \code{data.bper.long}. The element
  \code{metadata} is a list of seven named vectors that include all the same
  information at the top of each .pt file. The element \code{data.isbm} is a
  data frame of the ISBM index data in the same format as found in the .pt
  file. The element \code{data.isbm.long} is a data frame and comprises the
  same ISBM index data reshaped into the more useful 'long' format. The age
  data are exluded from \code{data.isbm.long} but can be found in
  \code{data.isbm.ages}, which is a data frame in long format. The element
  \code{data.bper} is data frame of the base perior exploitation rate data
  (BPER), in the same format as found in the .pt file. The element
  \code{data.bper.long} is the same BPER data but in long format.
}
\examples{
\dontrun{
data.path <- "../data/isbm"
pt.filenames <- list.files(data.path, pattern = "*step11*.csv$")
pt.filepaths <- paste(data.path, pt.filenames, sep="/")
pt.list <- readPTfromstep11csv(pt.filepaths)
}
}
