% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spfi_functions.R
\name{readStockData}
\alias{readStockData}
\title{(SPFI) Read CTC stock file (stocfile.stf).}
\usage{
readStockData(filename = "stocfile.stf")
}
\arguments{
\item{filename}{A string of length one. The default value is "stocfile.stf".}
}
\value{
A list of four elements. The first element, \code{stockmeta}, holds
  the meta data found in the first five lines of the stock file. The second
  element, \code{SPFIFlag}, is a data frame defining what stock-age
  combinations will be included in the analysis. The third element,
  \code{SPFIFlag.long}, is the same as \code{SPFIFlag}, but reshaped to a
  long format. The fourth element, \code{stocks.df}, is the table of stocks
  (found in the stock file).
}
\examples{
\dontrun{
data.stock <- readStockData('STOCFILE.STF')
}
}

