% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTHECCC.R
\name{readTHECCC}
\alias{readTHECCC}
\title{Read the model output CCC raw file ("THE" CCC file)}
\usage{
readTHECCC(filename, df.rows = 25, agegroups.n = 4)
}
\arguments{
\item{agegroups.n}{}
}
\value{
A list comprising as many lists as data sets in the CCC file.
}
\description{
This is a draft function. Variable naming isn't useful. But data
  structure is easily manipulated in R.
}
