% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spfi_functions.R
\name{reshapeHRJ}
\alias{reshapeHRJ}
\title{Reshape wide formatted HRJ file to long format.}
\usage{
reshapeHRJ(hrj.list, data.stock, fishery.def.df = NULL,
  jurisdiction.df = NULL)
}
\arguments{
\item{hrj.list}{A list. Output from \code{\link{readHRJAccessDatabase}}.}

\item{data.stock}{A list. Output from \code{\link{readStockData}}.}
}
\value{
A list comprising two elements, which are both data frames of the HRJ
  data tables 'B' & 'C' in long format.
}
\examples{
\dontrun{
# reading 32 bit mdb files requires using 32bit R
hrj.list.wide <- readHRJAccessDatabase("HRJ_database 2016b.mdb")
hrj.list.long <- reshapeHRJ(hrj.list.wide, data.stock)
}

}

