% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spfi_functions.R
\name{reshapeHRJtolong}
\alias{reshapeHRJtolong}
\title{(SPFI) Reshape wide formatted HRJ file to long format.}
\usage{
reshapeHRJtolong(hrj.list, data.stock, fishery.def.df = NULL,
  jurisdiction.df = NULL)
}
\arguments{
\item{hrj.list}{A list comprising data frames of HRJ data in wide format with
the same structure as seen in the output from
\code{\link{readHRJAccessDatabase}}.}

\item{data.stock}{A list. Output from \code{\link{readStockData}}.}

\item{fishery.def.df}{A data frame. The default is NULL, and if so the
\code{\link{fishery.def}} data are utilized.}

\item{jurisdiction.df}{A data frame. The default is NULL, and if so the
\code{\link{jurisdiction}} data are utilized.}
}
\value{
A list comprising three elements, all are data frames of the HRJ data
  tables 'B', 'C', and 'workingdata', in long format.
}
\description{
In addition to reshaping the wide tables to a long format, age
  and return year are calculated and the columns included. The data are
  evaluated for completeness of return year data, by stock. A (Boolean value)
  column named \code{return.year.complete} is included. A third data frame
  named "workingdata" is created. This consists of data from the 'c' table,
  which are revised with 'b' data based on the rules outlined in the details
  section below. The three data frames are also merged with the fishery
  definition \code{\link{fishery.def}} and jurisdiction
  \code{\link{jurisdiction}} data frames, which add additional meta data
  columns.
}
\details{
The data frame 'workingdata' represents a combination of data from
  the 'C' and 'B' data frames. The 'C' data frame is used to populate the
  'workingdata' data frame. The maximum number of age classes allowed to be
  missing, by stock and return year, is one. If more than one age class is
  absent, the 'C' value is replaced by its equivalent "B" value.

  The fishery definition \code{\link{fishery.def}} and jurisdiction
  \code{\link{jurisdiction}} data can be accessed by looking to their help
  file example documentation. If either data frame needs modification, the
  user can create new data frames with structure matching the two data frames
  included. The revised data frames can then be used as arguments to
  \code{reshapeHRJtolong}.
}
\examples{
\dontrun{
# reading 32 bit mdb files requires using 32bit R
hrj.list.wide <- readHRJAccessDatabase("HRJ_database 2016b.mdb")
hrj.list.long <- reshapeHRJtolong(hrj.list.wide, data.stock)
}

}

