% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spfi_functions.R
\name{reshapeHRJtowide}
\alias{reshapeHRJtowide}
\title{(SPFI) Reshape HRJ data frame from long to wide format for export to database.}
\usage{
reshapeHRJtowide(hrj.list)
}
\arguments{
\item{hrj.list}{A list with at least one named data frame, in long format. This function was intended to prepare the 'workingdata' data frame for export, but could also be applied to the 'C' and 'B' data frames (long format).}
}
\value{
A data frame in wide format (same structure as HRJ database tables).
}
\examples{
\dontrun{
hrj.list.wide <- reshapeHRJtowide(hrj.list.long)
}

}

