% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spfi_functions.R
\name{reshapeWorkingData}
\alias{reshapeWorkingData}
\title{Reshape HRJ data frame from long to wide format for export to database.}
\usage{
reshapeWorkingData(workingdata)
}
\arguments{
\item{workingdata}{A data frame, in long format and typically form the HRJ list.}
}
\value{
A data frame in wide format (same structure as HRJ database tables).
}
\description{
Reshape HRJ data frame from long to wide format for export to database.
}
\examples{
\dontrun{
workdingdata.wide <- reshapeWorkingData(hrj.list.long$working.data)
}

}

