% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{stocks.key}
\alias{stocks.key}
\title{Data defining each stock.}
\format{A data frame with 78 rows and 9 variables:
\describe{
  \item{StockNumber}{The integer number of each stock.}
  \item{Stock}{The three letter term for each stock.}
  \item{Description}{The full name for each stock.}
   ... }}
\source{
\url{http://www.psc.org/}
}
\usage{
stocks.key
}
\description{
A dataset containing the stock number and three letter definition.
}
\examples{
data(stocks.key)
#or
data("stocks.key")
str(stocks.key)
}
\keyword{datasets}
