% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chinookmodelfunctions.R
\name{sumPMs}
\alias{sumPMs}
\title{(Calibration Performance) Performance Measure Summation}
\usage{
sumPMs(dat)
}
\arguments{
\item{dat}{A dataframe, see details.}
}
\value{
A dataframe of the summed PM values (across ages), grouped by calibration model.
The column \code{allages} is the sum of RMSE values across stocks, for all \code{agegroup}'s
excludeing "totalabundance". The column \code{totalabundance} is the sum of RMSE
values across stocks, for only the "totalabundance" agegroup.
}
\description{
Cumulative sums of PM values for model ranking.
}
\details{
The arguement \code{dat} is generally the outpt from \code{\link{calcPMs}}.
}

