% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRJfunctions.R
\name{updateStockByName}
\alias{updateStockByName}
\title{(HRJ) Add stock number to HRJ data frame based on its three letter stock name.}
\usage{
updateStockByName(df, stockdat)
}
\arguments{
\item{df}{A data frame with column "stock.name" containing the three letter
stock name.}

\item{stockdat}{A data frame with columns "Stock.Number" and "StockID".
"StockID" contains the three letter stock name. This data frame is obtained
using \code{\link{readStockData}}.}
}
\value{
The function returns a data frame, same as used for the first
  argument, but with a "stock.number" column.
}
\examples{
\dontrun{
data.stock <- readStockData( 'STOCFILE.STF')
hrj.list <- readHRJtext(filepath)
hrj.list$hrj.cwt.list <- lapply(hrj.list$hrj.cwt.list, updateStockByName, data.stock$stocks.df)
}
}
