% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp_functions.R
\name{write_fpa}
\alias{write_fpa}
\title{Write FPA text file}
\usage{
write_fpa(fpdat, spfi, stocks, comment = NA, filename = NA)
}
\arguments{
\item{fpdat}{A data frame, equivalent structure to the first element of the
list that is produced by \code{\link{calc_fp}}}

\item{spfi}{A data frame. Annual estimates of the SPFI. Must include columns named
'return.year' and 'S.y' (the SPFI).}

\item{stocks}{A data frame with the first column identifying the three letter
stock name and the second column being the stock number. Column names are
not necessary.}

\item{comment}{A character vector of length one. The comment to be included in line 1
of the FPA file.}

\item{filename}{A character vector of length one. Required filename for the
output FPA file.}
}
\value{
This function only writes the FPA text file. Nothing is returned.
}
\description{
Write FPA text file
}
\examples{
\dontrun{
dat.fp <- calc_fp(...)
write_fpa(dat.fp[[1]], spfi, stocks, comment="wcvi", "fpa_results.fpa", )
}
}
