% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chinookmodelfunctions.R
\name{writeCalibrationTable4}
\alias{writeCalibrationTable4}
\title{(Calibration Performance) Write Table4}
\usage{
writeCalibrationTable4(metrics, ranking.method, results.path, ...)
}
\arguments{
\item{results.path}{}
}
\description{
Write table 4 as a text file.
}
