% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chinookmodelfunctions.R
\name{writeCalibrationTable5}
\alias{writeCalibrationTable5}
\title{(Calibration Performance)}
\usage{
writeCalibrationTable5(metrics, ranking.method, results.path, ...)
}
\arguments{
\item{...}{}
}
