% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newAWKfunctions.R
\name{writeHRJaccess}
\alias{writeHRJaccess}
\title{Write HRJ "B" & "C" tables to MS Access database.}
\usage{
writeHRJaccess(hrj, filename)
}
\arguments{
\item{hrj}{A list usually comprising of two data frames, which are the 'b'
and 'c' HRJ tables in wide format with fields exactly matching those
defined in the MS Access data base.}

\item{filename}{A character string of length one. The MS Access filename.}
}
\description{
The Access data base must already be created, but can be empty.
  If there are tables with the same names as the data frames, then they will
  be over-written.
}

