% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRJfunctions.R
\name{writeHRJAccessDatabase}
\alias{writeHRJAccessDatabase}
\title{(HRJ) Write HRJ "B" & "C" tables to MS Access database.}
\usage{
writeHRJAccessDatabase(hrj, extraTables = NA, filename)
}
\arguments{
\item{hrj}{A list usually comprising of two data frames, which are the 'b'
and 'c' HRJ tables in wide format with fields exactly matching those
defined in the MS Access data base.}

\item{extraTables}{A list comprising of one or more named elements that must
all be data frames. Each data frame will be written to a table that has the
same name as the data frame. If the list argument is not supplied, only a readme
table is created with the creation date in its contents.}

\item{filename}{A character string of length one. The MS Access filename.}
}
\value{
A MS database is written. Nothing is returned to R.
}
\description{
The Access data base must already be created, but can be empty.
  If there are tables with the same names as the data frames, then they will
  be over-written.
}
\examples{
\dontrun{
hrj.list <- readHRJtext(filepath)
hrj.list$hrj.cwt.list <- lapply(hrj.list$hrj.cwt.list,updateStockByName, data.stock$stocks.df)
writeHRJAccessDatabase(hrj = hrj.list$hrj.cwt.list, filename = 'test.accdb')

#to add the "workingdata" table (which has C data, updated with B data):
hrj.list.long <- reshapeHRJtolong(hrj.list$hrj.cwt.list, data.stock)
workdingdata.wide <- reshapeHRJtowide(hrj.list.long$workingdata)
writeHRJAccessDatabase(hrj = list(workingdata= workdingdata.wide), filename = 'test.accdb')
}
}
