% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newAWKfunctions.R
\name{writeHRJcsv}
\alias{writeHRJcsv}
\title{Write HRJ "B" & "C" tables to csv (text) files.}
\usage{
writeHRJcsv(hrj)
}
\arguments{
\item{hrj}{A list usually comprising of two data frames, which are the 'b'
and 'c' HRJ tables in wide format with fields exactly matching those
defined in the MS Access data base.}
}
\value{
Nothing is returned. But one csv file is written for each data frame
  in the HRJ list. The csv filename is the same as the data frame name.
}
\description{
This function is created if there are problems writing direct to
  the MS Access database. The user can create the csv files then import them
  from within Access.
}

