% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chinookmodelfunctions.R
\name{writeModelListBuilder}
\alias{writeModelListBuilder}
\title{(Calibration Performance) Build, save, and open an R script to
  simplify creation of \code{model.list} object.}
\usage{
writeModelListBuilder()
}
\value{
Opens an R script.
}
\description{
This creates and opens a script named "ModelListBuilder.R". This
  is a template for the user to work with when doing performance analysis of
  the calibration models. This is intended to help the user automatically
  build the \code{model.list} object, which is used as an argument to many
  functions. Some object values will need updating (for example the paths).
}
\examples{
writeModelListBuilder()
}
