% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chinookmodelfunctions.R
\name{writeScriptCalibrationTester}
\alias{writeScriptCalibrationTester}
\title{(Calibration Performance) Build, save, and open an R script to help
  run calibration model performance analysis.}
\usage{
writeScriptCalibrationTester()
}
\value{
Opens an R script that includes the template of functions to evaluate
  CCC performance.
}
\description{
This creates and opens a script named "CalibrationTester.R".
  This is a template for the user to work with when doing performance
  analysis of the calibration models. This is intended to help the user
  understand the work flow. However, due to file path differences, the script
  may not work as is. Some object values will need updating (for example the
  paths).
}
\examples{
writeScriptCalibrationTester()
}
