% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRJfunctions.R
\name{writeScriptHRJ}
\alias{writeScriptHRJ}
\title{(SPFI) Build, save, and open an R script to help manage HRJ data.}
\usage{
writeScriptHRJ()
}
\value{
Opens an R script that includes the template of functions.
}
\description{
This creates and opens a script named "hrj_script.R". This is a
  template for the user to work with when managing HRJ data (import,
  manipulate, export). This is intended to help the user understand the work
  flow and due to file path differences, is unlikely to work as is. Some
  object values will need updating (for example the datapath).
}
\examples{
writeScriptHRJ()
}
