% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isbm_functions.R
\name{writeScriptISBM}
\alias{writeScriptISBM}
\title{(ISBM/CYER) Build, save, and open an R script to help compare ISBM &
  CYER indices.}
\usage{
writeScriptISBM()
}
\value{
Opens an R script that includes the functions to compare indices.
}
\description{
This creates and opens a script named "isbm_script.R". This is a
  template for the user to work with when doing comparison of ISBM and CYER
  indices. This is intended to help the user understand the work flow and due
  to file path differences, is unlikely to work as is. Some object values
  will need updating (for example the data path).
}
\examples{
writeScriptISBM()
}
