% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chinookmodelfunctions.R
\name{writeTable1}
\alias{writeTable1}
\title{(Calibration Performance) Write Table1 To csv}
\usage{
writeTable1(data.combined, results.path = ".")
}
\arguments{
\item{data.combined}{}
}
\description{
Write table 1 as a csv, by stock, calibration, & data.type.
}

