% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chinookmodelfunctions.R
\name{writeTable3}
\alias{writeTable3}
\title{(Calibration Performance) Write Table3}
\usage{
writeTable3(metrics, ranking, results.path, ...)
}
\arguments{
\item{results.path}{}
}
\description{
Write table 3 as a text file.
}

