% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chinookmodelfunctions.R
\name{writeTable4}
\alias{writeTable4}
\title{Write Table4}
\usage{
writeTable4(metrics, ranking, results.path, ...)
}
\arguments{
\item{results.path}{}
}
\description{
Write table 4 as a text file.
}

