% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chinookmodelfunctions.R
\name{writeTableOfDifferences}
\alias{writeTableOfDifferences}
\title{(Calibration Performance)}
\usage{
writeTableOfDifferences(metrics, ranking, results.path,
  tabletype = c("table3", "table4"), ...)
}
\arguments{
\item{...}{}
}

