% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chinookmodelfunctions.R
\name{buildmodel.list}
\alias{buildmodel.list}
\title{(Calibration Performance) Automatic Creation Of Input List}
\usage{
buildmodel.list(stock.names = "all", commonstocks = FALSE,
  stockmap.pathname, data.path.vec, stocks.key.pathname.vec, grouping.year,
  age.structure, totalabundance, data.type = c("escapement", "terminalrun"),
  results.path = NA, groupingby = c("stock", "agegroup"),
  ranking = c("ordinal", "interpolated"))
}
\arguments{
\item{stock.names}{A vector of the three letter stock name acronyms, or "all" for all stocks.}

\item{results.path}{}
}
\description{
Constructs list of arguements to be utilized by numerous functions within PBSctc.
}

