% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chinookmodelfunctions.R
\name{buildmodel.list}
\alias{buildmodel.list}
\title{(Calibration Performance) Automatic Creation Of argument List}
\usage{
buildmodel.list(stock.names = "all", commonstocks = FALSE,
  stockmap.pathname, data.path.vec, stocks.key.pathname.vec,
  startingyear = NA, finalyear = NULL, grouping.year, age.structure,
  totalabundance, data.type = c("escapement", "terminalrun"),
  results.path = NA, groupingby = c("stock", "agegroup"),
  ranking.method = c("ordinal", "interpolated"))
}
\arguments{
\item{stock.names}{A vector of the three letter stock name acronyms, or "all"
for all stocks.}

\item{commonstocks}{A Boolean. Whether to compare results based only on
common stocks across calibrations. If only running one model keep as FALSE.
The default is FALSE}

\item{stockmap.pathname}{A string. The path to the stockmap file.}

\item{data.path.vec}{A string. The path to the ccc and fcs files.}

\item{stocks.key.pathname.vec}{A string. The path to the stock key file.}

\item{age.structure}{A Boolean. Include analyses based on age structure.}

\item{totalabundance}{A Boolean. Include analyses for stocks having only
total abundance data. year totals?}

\item{data.type}{A string of length one or two. The values can be
"escapement" and "terminalrun".}

\item{results.path}{A character vector of length one. The absolute path or
path relative to the working directory where a new directory named
"results" will be written (if not already present). All output files will
be written here. Default is the current working directory.}
}
\value{
A list, to be used as the argument to many functions.
}
\description{
Constructs list of arguments to be utilized by numerous
  functions within ctctools.
}
\examples{
\dontrun{
commonstocks <- FALSE
stock.names <- 'all'
stockmap.pathname <- "../data/Old-New_Stock_Mapping_20160916.csv"
data.path.vec <- c("../data/2016_CI_test")
stocks.key.pathname.vec <- c("../data/calibration2017/CLB17bb/stocks.oldnames.csv")
# this can be "brood.year" or "return.year":
grouping.year <- "return.year"
age.structure <- TRUE
totalabundance <- TRUE
data.type <- c("escapement", "terminalrun")
samplesize.min <- 10
results.path <- "../data/calibration2017"
ranking.method <- c('ordinal', "interpolated")
doPlots <- TRUE  # TRUE OR FALSE. ALL CAPS
savepng <-  TRUE  # TRUE OR FALSE. ALL CAPS
model.list <- buildmodel.list(commonstocks = commonstocks, stockmap.pathname = stockmap.pathname, data.path.vec = data.path.vec, stocks.key.pathname.vec = stocks.key.pathname.vec, grouping.year = grouping.year, age.structure = age.structure, totalabundance =totalabundance, data.type=data.type, results.path = results.path, stock.names = stock.names, groupingby=c( "agegroup"), ranking.method = ranking.method)
}
\dontrun{
model.list <- buildmodel.list(commonstocks = TRUE, stockmap.pathname = stockmap.pathname, data.path.vec = data.path.vec, stocks.key.pathname.vec = stocks.key.pathname.vec, grouping.year = grouping.year, age.structure = age.structure, totalabundance =totalabundance, data.type=data.type, results.path = results.path, stock.names = stock.names, groupingby=c( 'agegroup'), ranking = ranking)
}
}
