% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spfi_functions.R
\name{calc_d.tsa}
\alias{calc_d.tsa}
\title{(SPFI) Calculate the distribution parameters grouped by fishery,
  stock, and age.}
\usage{
calc_d.tsa(r.tsa.sum, n.ysa, hcwt.ty = NULL, standardize.bol = FALSE)
}
\arguments{
\item{r.tsa.sum}{Output from \code{\link{calc_tsa.sum}}.}

\item{n.ysa}{Synonymous with \code{CWTPop} in VB or:
\code{hrj.df[hrj.df$data.type=="Pop" & hrj.df$fishery.index == 1,]}}

\item{hcwt.ty}{Output from \code{\link{calc_hcwt.ty}}.}

\item{standardize.bol}{A Boolean, default=FALSE.}
}
\value{
A data frame of the distribution parameter estimates grouped by
  fishery stratum, stock, and age.
}
\description{
This is equivalent to equation 1 in the draft SPFI document.
}
\details{
The initial call of this function will have no data for
  \code{hcwt.ty}. If this is true, then the values are set to 0.01 as is done
  in the Visual Basic.
}
\examples{
\dontrun{ d.tsa <- calc_d.tsa(r.tsa.sum = r.tsa.sum, n.ysa = cwtpop,
standardize.bol = TRUE) }

}

