% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp_functions.R
\name{calc_fp}
\alias{calc_fp}
\title{(FP) Calculate fisheries policy (FP) scaler.}
\usage{
calc_fp(dat.er.long, dat.mdl.long, dat.spfi)
}
\arguments{
\item{dat.er.long}{Data frame returned by \code{\link{read_stkfile}}.}

\item{dat.mdl.long}{Data frame returned by \code{\link{read_mdl}}.}

\item{dat.spfi}{Data frame in long format with same structure as the element
\code{S.ty}, which is output of \code{\link{calc_SPFI}}.}
}
\value{
A data frame comprising the FP values.
}
\examples{
\dontrun{
dat.fp <- calc_fp(dat.er.long = dat.er.long.sub,
  dat.mdl.long = dat.mdl.long.sub, dat.spfi = dat.spfi.long)

}
}
