% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spfi_functions.R
\name{calc_hcwt.ty}
\alias{calc_hcwt.ty}
\title{(SPFI) Calculate the CWT harvest rate parameters grouped by fishery,
  and year.}
\usage{
calc_hcwt.ty(r.ty.sum, d.tsa, n.ysa)
}
\arguments{
\item{r.ty.sum}{Output from \code{\link{calc_ty.sum}}.}

\item{d.tsa}{Output from \code{\link{calc_d.tsa}}.}

\item{n.ysa}{Synonymous with \code{CWTPop} in VB or:
\code{hrj.df[hrj.df$data.type=="Pop" & hrj.df$fishery.index == 1,]}}
}
\description{
This is equivalent to equation 2 in the draft SPFI document.
}

