% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spfi_functions.R
\name{calc_N.y}
\alias{calc_N.y}
\title{(SPFI) Calculate the total yearly abundance across fishery strata.}
\usage{
calc_N.y(N.ty)
}
\arguments{
\item{N.ty}{Output from \code{calc_N.ty}.}
}
\description{
This is equivalent to equation 5 in the draft SPFI document.
}

