% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chinookmodelfunctions.R
\name{calcMPEfreq}
\alias{calcMPEfreq}
\title{(Calibration Performance) Frequency counts by MPE bin ranges.}
\usage{
calcMPEfreq(metrics.arg, results.path = ".", mpe.range.vec = c("abs", "neg",
  "pos"), ...)
}
\arguments{
\item{metrics.arg}{A data frame. Output of \code{\link{calcPMs}}.}

\item{results.path}{A character vector of length one. The absolute path or
path relative to the working directory where a new directory named
"results" will be written (if not already present). All output files will
be written here. Default is the current working directory.}

\item{mpe.range.vec}{A vector}

\item{...}{Optional arguments.}
}
\value{
A csv file for each value in the argument \code{mpe.range.vec}.
}
\examples{
\dontrun{
calcMPEfreq(metrics, results.path = model.list$results.path, mpe.range.vec = c('pos', 'neg', 'abs'))
}

}
