% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chinookmodelfunctions.R
\name{calcRanks}
\alias{calcRanks}
\title{Rank Calculation}
\usage{
calcRanks(dat, columnToRank, rank.method = c("ordinal", "interpolated"),
  abs.value = FALSE)
}
\arguments{
\item{dat}{A dataframe}

\item{columnToRank}{A vector of integers defining what columns to rank (independently)}

\item{rank.method}{A character defining the ranking method to apply.}
}
\value{
A dataframe comprising all columns of the argument: \code{dat} and one
additional column of ranks for each column column declared in \code{columnToRank}.
The name of each rank column is a combination of the data column name and ".rank".
}
\description{
Calculate rank of data.
}

