% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRJfunctions.R
\name{checkMissingfiles}
\alias{checkMissingfiles}
\title{(HRJ) Check for matching B & C file names in the vector of file names.}
\usage{
checkMissingfiles(filepath)
}
\arguments{
\item{filepath}{A character vector, which can have length greater than one.
Each element of the vector is the path and filename for each HRJ file to be
read in.}
}
\value{
No object is returned. If files are included in the vector but lack a
  matching alternate data, then those are printed to the console.
}
\description{
This tests for that every file named in the filepath, whether
  'b' or 'c' data, the alternate data ('c' or 'b') by stock, has also been
  named.
}

