% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctctools.R
\docType{package}
\name{ctctools}
\alias{ctctools}
\alias{ctctools-package}
\title{ctctools: Analytical tools for the Chinook Technical Committee.}
\description{
The package comprises functions from four projects that have moderate overlap:
\enumerate{
\item Performance analyses of calibration results,
\item HRJ data file import & manipulation (revision to the AWK code),
\item HRJ data plotting (Not currently included),
\item Calculation of the Stratified Proportional Fishery Index (SPFI).
}
}
\section{}{

}

