% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chinookmodelfunctions.R
\name{plotPM}
\alias{plotPM}
\title{(Calibration Performance) Lattice plot to compare PMs across
  calibration models}
\usage{
plotPM(tableofdifferences, results.path = ".", savepng = FALSE)
}
\arguments{
\item{tableofdifferences}{A data frame, Output of
\code{\link{importTableOfDifferences}}}

\item{results.path}{A character vector of length one. The absolute path or
path relative to the working directory where a new directory named
"results" will be written (if not already present). All output files will
be written here. Default is the current working directory.}

\item{savepng}{A Boolean. Save output to a png file. Default is FALSE.}
}
\value{
Lattice plot of the data found in the "Table of differences" files.
}
\description{
(Calibration Performance) Lattice plot to compare PMs across
  calibration models
}
\examples{
\dontrun{
plotPM(tableofdifferences = tableofdifferences,
 results.path = model.list$results.pat,savepng=TRUE )
}
}
