% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spfi_functions.R
\name{readCatchData}
\alias{readCatchData}
\title{(SPFI) Read the CTC catch data files (*.cat).}
\usage{
readCatchData(filename, strLocation = c("seak", "nbc", "wcvi"))
}
\arguments{
\item{filename}{A string of length one, defining the name of the catch file.
If more than one file name is included, only the first is used.}

\item{strLocation}{A string of length one, defining the AABM location
("seak", "nbc", "wcvi").}
}
\value{
A list of six elements. The sixth element is a data frame of the
  catch data.
}
\description{
(SPFI) Read the CTC catch data files (*.cat).
}
\details{
Needs details.
}
\examples{
\dontrun{
readCatchData("wcvi7914.cat", strLocation = "wcvi")
}
}
