% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chinookmodelfunctions.R
\name{readCCC}
\alias{readCCC}
\title{(Calibration Performance) Import CCC Files}
\usage{
readCCC(filepath, data.types = c("AEQ", "cohort", "termrun", "escape"),
  stocks.key = NA, finalyear = 9999)
}
\arguments{
\item{filepath}{A character vector of the CCC files to be imported. See details}

\item{data.types}{See details}

\item{stocks.key}{See details}

\item{finalyear}{An integer. The final (4 digit) year to be included from all series
imported. Default is 9999, meaning all years before 9999.}
}
\value{
A list, with one element per CCC (calibration) file.
}
\description{
Import one or more CCC files into a list.
}
\details{
Need details.
}
\examples{
\dontrun{
### read in CCC files ###
filename <- list.files(data.pathname, pattern = "CCC")
filepath <- paste(data.pathname,  filename, sep='/')
ccc.list <- sapply(filepath, readCCC, USE.NAMES = FALSE)
}
}
