% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chinookmodelfunctions.R
\name{readFCS}
\alias{readFCS}
\title{(Calibration Performance) Import FCS Files}
\usage{
readFCS(filepath, first.stockline = 3, stocks.key = NA)
}
\arguments{
\item{filepath}{A character vector of length one. The name of the FCS file to
import.}

\item{first.stockline}{An integer, row number in FCS file where first stock
metadata commences. See details.}

\item{stocks.key}{See details}
}
\value{
A list, with two elements. The FCS file data in wide and long
  formats, respectively.
}
\description{
Import one or more FCS files into a list.
}
\details{
Need details.
}
\examples{
\dontrun{
### read in FCS files ###
fcs.files <- list.files(data.pathname, pattern = "\\\\.FCS")
filename <- fcs.files[grep("OCN", x = fcs.files)]
filepath <- paste(data.pathname, filename, sep='/')
fcs.list <- readFCS(filepath)
}
}
