% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chinookmodelfunctions.R
\name{readFCS}
\alias{readFCS}
\title{Import FCS Files}
\usage{
readFCS(filepath, first.stockline = 3, stocks.key = NA)
}
\arguments{
\item{filepath}{A character vector of the CCC files to be imported. See details}

\item{first.stockline}{An integer, row number in FCS file where first stock
metadata commences. See details.}

\item{stocks.key}{See details}
}
\value{
A list, with one element per CCC (calibration) file.
}
\description{
Import one or more FCS files into a list.
}
\details{
Need details.
}

