% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRJfunctions.R
\name{reshapeHRJtowide}
\alias{reshapeHRJtowide}
\title{(HRJ) Reshape HRJ data frame from long to wide format for export to database.}
\usage{
reshapeHRJtowide(hrj.list)
}
\arguments{
\item{hrj.list}{A list with at least one named data frame, in long format.}
}
\value{
A list of with the same number of data frames, in wide format (same structure as HRJ database tables), as was found in the list included in the function argument.
}
\description{
This function was intended to prepare the 'workingdata' data frame for export, but could also be applied to the 'C' and 'B' data frames (long format).
}
\examples{
\dontrun{
hrj.list.wide <- reshapeHRJtowide(hrj.list.long)
}
}
