% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chinookmodelfunctions.R
\name{tabulateMetrics}
\alias{tabulateMetrics}
\title{(Calibration Performance) Tabulate Ranking}
\usage{
tabulateMetrics(metrics, groupingby, ...)
}
\arguments{
\item{metrics}{A \code{list}. See Details}

\item{...}{Further arguments. Currently limited to \code{rank.method}, which
is sent to \code{\link{calcRanks}}.}
}
\value{
A list, with length equal to the number of ranking methods
chosen.
}
\description{
Tabulate Ranking Of Calibation Models
}
\details{
This function calls \code{\link{calcRanks}} and organizes the ouput for
comparing performance metric specific ranking of calibration models.
The argument \code{metrics} is a list and is to be taken from the output of
\code{\link{calcPMs}}.
}

