% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chinookmodelfunctions.R
\name{writeCalibrationTable3}
\alias{writeCalibrationTable3}
\title{(Calibration Performance) Write Table3}
\usage{
writeCalibrationTable3(metrics, ranking.method, results.path, ...)
}
\arguments{
\item{results.path}{}
}
\description{
Write table 3 as a text file.
}
