% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRJfunctions.R
\name{writeHRJAccessDatabase}
\alias{writeHRJAccessDatabase}
\title{(HRJ) Write HRJ "B" & "C" tables to MS Access database.}
\usage{
writeHRJAccessDatabase(hrj, filename)
}
\arguments{
\item{hrj}{A list usually comprising of two data frames, which are the 'b'
and 'c' HRJ tables in wide format with fields exactly matching those
defined in the MS Access data base.}

\item{filename}{A character string of length one. The MS Access filename.}
}
\description{
The Access data base must already be created, but can be empty.
  If there are tables with the same names as the data frames, then they will
  be over-written.
}
\examples{
\dontrun{
hrj.list <- readHRJtext(filepath)
hrj.list$hrj.cwt.list <- lapply(hrj.list$hrj.cwt.list,updateStockByName, data.stock$stocks.df)
writeHRJAccessDatabase(hrj = hrj.list$hrj.cwt.list, filename = 'test.accdb')

#to add the "workingdata" table (which has C data, updated with B data):
hrj.list.long <- reshapeHRJtolong(hrj.list$hrj.cwt.list, data.stock)
workdingdata.wide <- reshapeHRJtowide(hrj.list.long$workingdata)
writeHRJAccessDatabase(hrj = list(workingdata= workdingdata.wide), filename = 'test.accdb')
}
}

