% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRJfunctions.R
\name{writeHRJcsv}
\alias{writeHRJcsv}
\title{(HRJ) Write HRJ "B" & "C" tables to csv (text) files.}
\usage{
writeHRJcsv(hrj)
}
\arguments{
\item{hrj}{A list usually comprising of two data frames, which are the 'b'
and 'c' HRJ tables in wide format with fields exactly matching those
defined in the MS Access data base.}
}
\value{
Nothing is returned. But one csv file is written for each data frame
  in the HRJ list. The csv filename is the same as the data frame name.
}
\description{
This function has been created in case there are future problems
  writing direct to the MS Access database. The user can create the csv files
  then import them from within Access.
}
\examples{
\dontrun{
hrj.list <- readHRJtext(filepath)
hrj.list$hrj.cwt.list <- lapply(hrj.list$hrj.cwt.list,updateStockByName, data.stock$stocks.df)
writeHRJcsv(hrj = hrj.list$hrj.cwt.list)
}
}

